Group tests to reduce time spent for testing
