--- component/src/test/java/org/ballerinalang/kafka/nativeimpl/producer/KafkaProducerGetPartitionTest.java
@@ -1,94 +1,0 @@
-/*
-*   Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-
-package org.ballerinalang.kafka.nativeimpl.producer;
-
-import io.debezium.kafka.KafkaCluster;
-import io.debezium.util.Testing;
-import org.ballerinalang.launcher.util.BCompileUtil;
-import org.ballerinalang.launcher.util.BRunUtil;
-import org.ballerinalang.launcher.util.CompileResult;
-import org.ballerinalang.model.values.BString;
-import org.ballerinalang.model.values.BValue;
-import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.Properties;
-
-/**
- * Test cases for ballerina.net.kafka producer connector/partition retrieval.
- */
-@Test(singleThreaded = true)
-public class KafkaProducerGetPartitionTest {
-    private CompileResult result;
-    private static File dataDir;
-    protected static KafkaCluster kafkaCluster;
-
-    @BeforeClass
-    public void setup() throws IOException {
-        result = BCompileUtil.compile("producer/kafka_producer_partition_retrieval.bal");
-        Properties prop = new Properties();
-        kafkaCluster = kafkaCluster().deleteDataPriorToStartup(true)
-                .deleteDataUponShutdown(true).withKafkaConfiguration(prop).addBrokers(1).startup();
-        kafkaCluster.createTopic("test", 2, 1);
-        kafkaCluster.createTopic("test_2", 5, 1);
-    }
-
-    @Test(description = "Test producer topic partition retrieval")
-    public void testKafkaTopicPartitionRetrieval() {
-        BValue[] inputBValues = {new BString("test")};
-        BValue[] returnBValues = BRunUtil.invoke(result, "funcTestPartitionInfoRetrieval", inputBValues);
-        Assert.assertEquals(returnBValues.length, 2);
-
-        inputBValues = new BValue[]{new BString("test_2")};
-        returnBValues = BRunUtil.invoke(result, "funcTestPartitionInfoRetrieval", inputBValues);
-        Assert.assertEquals(returnBValues.length, 5);
-
-        //negative test for the case where topic has not been created programmatically
-        inputBValues = new BValue[]{new BString("test_negative")};
-        returnBValues = BRunUtil.invoke(result, "funcTestPartitionInfoRetrieval", inputBValues);
-        Assert.assertEquals(returnBValues.length, 1);
-    }
-
-    @AfterClass
-    public void tearDown() {
-        if (kafkaCluster != null) {
-            kafkaCluster.shutdown();
-            kafkaCluster = null;
-            boolean delete = dataDir.delete();
-            // If files are still locked and a test fails: delete on exit to allow subsequent test execution
-            if (!delete) {
-                dataDir.deleteOnExit();
-            }
-        }
-    }
-
-    protected static KafkaCluster kafkaCluster() {
-        if (kafkaCluster != null) {
-            throw new IllegalStateException();
-        }
-        dataDir = Testing.Files.createTestingDirectory("cluster-kafka-producer-get-partitions-test");
-        kafkaCluster = new KafkaCluster().usingDirectory(dataDir).withPorts(2182, 9094);
-        return kafkaCluster;
-    }
-
-}
