+++ component/src/test/java/org/ballerinalang/kafka/nativeimpl/services/KafkaConsumerServiceTest.java
@@ -120,6 +120,24 @@ public class KafkaConsumerServiceTest {
         }
     }
 
+    @Test(description = "Test endpoint bind to a service")
+    public void testKafkaAdvancedService() {
+        compileResult = BCompileUtil.compileAndSetup("services/kafka_service_advanced.bal");
+        BServiceUtil.runService(compileResult);
+        BRunUtil.invokeStateful(compileResult, "funcKafkaProduce");
+
+        try {
+            await().atMost(10000, TimeUnit.MILLISECONDS).until(() -> {
+                BValue[] returnBValues = BRunUtil.invokeStateful(compileResult, "funcKafkaGetResultText");
+                Assert.assertEquals(returnBValues.length, 1);
+                Assert.assertTrue(returnBValues[0] instanceof BBoolean);
+                return ((BBoolean) returnBValues[0]).booleanValue();
+            });
+        } catch (Throwable e) {
+            Assert.fail(e.getMessage());
+        }
+    }
+
     @AfterClass
     public void tearDown() {
         if (kafkaCluster != null) {
