+++ component/src/test/resources/producer/kafka_producer_flush_records.bal
@@ -16,6 +16,8 @@
 
 import wso2/kafka;
 
+string topic = "producer-flush-records-test-topic";
+
 kafka:ProducerConfig producerConfigs = {
     bootstrapServers: "localhost:9094",
     clientID: "flush-records-producer",
