+++ pom.xml
@@ -23,7 +23,7 @@
     <groupId>org.ballerinalang</groupId>
     <artifactId>ballerina-kafka-parent</artifactId>
     <packaging>pom</packaging>
-    <version>0.972.1-SNAPSHOT</version>
+    <version>0.972.0-SNAPSHOT</version>
     <name>Ballerina - Kafka Parent</name>
 
     <scm>
@@ -229,20 +229,6 @@
                 <classifier>ballerina-binary-repo</classifier>
             </dependency>
             <dependency>
-                <groupId>org.ballerinax.docker</groupId>
-                <artifactId>docker-extension</artifactId>
-                <type>zip</type>
-                <classifier>ballerina-binary-repo</classifier>
-                <version>${ballerina.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.ballerinax.kubernetes</groupId>
-                <artifactId>kubernetes-extension</artifactId>
-                <type>zip</type>
-                <classifier>ballerina-binary-repo</classifier>
-                <version>${ballerina.version}</version>
-            </dependency>
-            <dependency>
                 <groupId>org.ballerinalang</groupId>
                 <artifactId>ballerina-core</artifactId>
                 <version>${ballerina.version}</version>
@@ -336,16 +322,6 @@
             <dependency>
                 <groupId>org.ballerinalang</groupId>
                 <artifactId>ballerina-http</artifactId>
-                <version>${ballerina.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.ballerinax.docker</groupId>
-                <artifactId>docker-extension</artifactId>
-                <version>${ballerina.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.ballerinax.kubernetes</groupId>
-                <artifactId>kubernetes-extension</artifactId>
                 <version>${ballerina.version}</version>
             </dependency>
             <dependency>
@@ -525,6 +501,7 @@
         <awaitility.version>3.0.0</awaitility.version>
         <project.scm.id>my-scm-server</project.scm.id>
         <ballerina.version>0.990.1-SNAPSHOT</ballerina.version>
+        <ballerinax.version>0.983.1-SNAPSHOT</ballerinax.version>
         <apache.kafka.version>2.0.0</apache.kafka.version>
         <org.testng.version>6.11</org.testng.version>
         <maven.findbugsplugin.version.exclude>findbugs-exclude.xml</maven.findbugsplugin.version.exclude>
